Create Ethan and Alya
Problem Statement



Ethan and Alya are working together on a simple stack-based program to search for elements in a fixed-size (30) stack implemented using an array. The program has two main operations: pushing elements onto the stack and searching for elements within the stack. They need to implement these operations efficiently.



Write a program to accomplish the above.



Note: This is a sample question asked during Infosys recruitment.

Input format :
The first line of input consists of an integer 'n', representing the number of elements to be pushed onto the stack.

The next line contains 'n' space-separated integers, each representing an element to be pushed onto the stack.

The next line of input consists of an integer s, representing the element to be searched for in the stack.

Output format :
The output prints one of the following outputs based on the input:



If the stack is full: "STACK is overflow"

If the stack is empty: "The STACK is empty"

If the searched element is found in the stack: "Element found"

If the searched element is not found in the stack: "Element not found"



Refer to the sample outputs for the exact format.

Code constraints :
1 <= n <= 30

-100 <= a[i] <= 100

-100 <= s <= 100

Sample test cases :
Input 1 :
5
1 2 3 4 5
3
Output 1 :
Element found
Input 2 :
5
-1 2 3 4 -5
-6
Output 2 :
Element not found
Input 3 :
0
Output 3 :
The STACK is empty
Input 4 :
31
1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31
67
Output 4 :
STACK is overflow
Element not found