mira
Problem Statement

Mira wants to create a program that allows her to manipulate a stack, perform push and pop operations, and display the elements in the stack. She would like a user-friendly interface that presents a menu with the following options:



Push Operation: Add an integer to the top of the stack.
Pop Operation: Remove and discard the integer from the top of the stack.
Display Stack: Display all the elements currently in the stack, from top to bottom.
Exit the Program: Terminate the program.


Note: This is a sample question asked in a Capgemini interview.

Input format :
The program expects Ganga to enter a choice (an integer) corresponding to the operation she wants to perform:

1: Push the integer value onto the stack.

2: Pop an integer from the stack.

3: Display the elements currently in the stack.

4: Exit the program.

Output format :
The output displays messages according to the chosen menu option and the status of the stack:



"Stack is empty. Cannot perform a pop operation." when attempting to pop from an empty stack.

"Stack is empty." when the stack is empty.

"Elements in the stack:..." followed by the elements in the stack when choosing option 3.

"Exiting the program." when choosing option 4.

"Invalid choice." when entering an invalid menu choice.



Refer to the sample output for the exact format.

Code constraints :
Choice = 1, 2, 3, 4

0 <= Stack elements <= 50

Sample test cases :
Input 1 :
1
5
1
10
1
25
2
3
4
Output 1 :
Elements in the stack: 10 5 
Exiting the program.
Input 2 :
1
5
3
4
Output 2 :
Elements in the stack: 5 
Exiting the program.
Input 3 :
1
2
1
7
2
2
3
4
Output 3 :
Stack is empty.
Exiting the program.
Input 4 :
1
4
2
2
3
4
Output 4 :
Stack is empty. Cannot perform pop operation.
Stack is empty.
Exiting the program.